# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RankingSavedStateResponse": {
        "properties": {
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          },
          "is_saved": {
            "title": "Is Saved",
            "type": "boolean"
          }
        },
        "required": [
          "content_id",
          "is_saved"
        ],
        "title": "RankingSavedStateResponse",
        "type": "object"
      },
...
      "SaveRankingRequest": {
        "properties": {
          "kind": {
            "enum": [
              "hacker_news",
              "reddit",
              "github"
            ],
            "title": "Kind",
            "type": "string"
          },
          "sort": {
            "default": "hot",
            "enum": [
              "hot",
              "rising",
              "top"
            ],
            "title": "Sort",
            "type": "string"
          },
          "subreddit": {
            "default": "MachineLearning",
            "maxLength": 100,
            "minLength": 2,
            "title": "Subreddit",
            "type": "string"
          },
          "time_filter": {
            "default": "week",
            "enum": [
              "day",
              "week",
              "month",
              "year"
            ],
            "title": "Time Filter",
            "type": "string"
          },
          "url": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "url"
        ],
        "title": "SaveRankingRequest",
        "type": "object"
      },
      "SavedItemResponse": {
        "properties": {
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Name"
          },
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "external_url": {
            "title": "External Url",
            "type": "string"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "is_saved": {
            "default": true,
            "title": "Is Saved",
            "type": "boolean"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "ranking_kind": {
            "anyOf": [
              {
                "enum": [
                  "hacker_news",
                  "reddit",
                  "github"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Kind"
          },
          "source_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Avatar Url"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "translated_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Title"
          },
          "translation_locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Locale"
          },
          "translation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Status"
          }
        },
        "required": [
          "content_id",
          "source_id",
          "source_name",
          "source_avatar_url",
          "source_kind",
          "title",
          "translated_title",
          "translation_locale",
          "translation_status",
          "excerpt",
          "external_url",
          "author_name",
          "published_at",
          "fetched_at",
          "thumbnail_url"
        ],
        "title": "SavedItemResponse",
        "type": "object"
      },
      "SavedPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "SavedPageResponse",
        "type": "object"
      },
      "SavedStateResponse": {
        "properties": {
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "is_saved": {
            "title": "Is Saved",
            "type": "boolean"
          }
        },
        "required": [
          "content_id",
          "is_saved"
        ],
        "title": "SavedStateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "ReaderServerToken": {
        "description": "Deployment connection token configured by the server administrator.",
        "in": "header",
        "name": "X-Reader-Server-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend for shared-source information aggregation.",
    "title": "Reader API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/saved": {
      "get": {
        "description": "Return the caller's saved items, newest saved first.",
        "operationId": "list_saved_content_v1_saved_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedItemResponse"
                  },
                  "title": "Response List Saved Content V1 Saved Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Saved Content",
        "tags": [
          "saved"
        ]
      }
    },
    "/v1/saved/page": {
      "get": {
        "description": "Return a stable saved-items page with SQL-level source de-duplication.",
        "operationId": "list_saved_content_page_v1_saved_page_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Saved Content Page",
        "tags": [
          "saved"
        ]
      }
    },
    "/v1/saved/ranking": {
      "get": {
        "operationId": "get_ranking_saved_state_v1_saved_ranking_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "maxLength": 8192,
              "minLength": 1,
              "title": "Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingSavedStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ranking Saved State",
        "tags": [
          "saved"
        ]
      },
      "put": {
        "operationId": "save_ranking_v1_saved_ranking_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRankingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Ranking",
        "tags": [
          "saved"
        ]
      }
    },
    "/v1/saved/{content_id}": {
      "delete": {
        "operationId": "unsave_content_v1_saved__content_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unsave Content",
        "tags": [
          "saved"
        ]
      },
      "put": {
        "operationId": "save_content_v1_saved__content_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Content",
        "tags": [
          "saved"
        ]
      }
    },
...
  }
}